backend/pyproject.toml
[project]
name = "mytdl"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shaade", email = "s.shaade77@gmail.com" }
]
dependencies = [
    "djangorestframework>=3.14.0",
    "markdown>=3.4.4",
    "django-filter>=23.3",
    "django>=4.2.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
